+++ index.html
@@ -51,12 +51,14 @@
                     <li>Our Phone number: 00463763468</li>
                     <li>Our Address: Streetname 22,6826 City, Country</li>
                 </ul>
-
             </section>
+
         </div>
+
         <footer>
             <h3>Copyright <span>...</span></h3>
         </footer>
+        
         <script type="module" src="index.js"></script>
     </body>
